requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
chromadb>=0.4.22
streamlit>=1.31.0
pypdf>=4.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.3.0
groq>=0.4.0
pandas>=2.0.0
tiktoken>=0.5.2
